Restructure folders
Restructure project layout to accommodate multiple apps in the same repo.
In the future, each app should have a readme in it.
tasks.py is an attempt to make building different apps easier.  The intent is that tasks.py will be used in tasks.json to make switching what app to build easier.  This script is not complete yet and there are issues with it.
tasks.json needs to be edited to change what app the tasks will run on.
launch.json needs to be edited depending on what app is going to be debugged.

bug 13038