apps: oob_demo: Add cell fw version and status
Add firmware version and status to cellular service.

Stop advertising when a central connects.
This will ensure another central cannot connect if the pinnacle is not yet connected to a sensor.
This also solves the bug where the pinnacle could not connect to a sensor if the central connected first.

bug 14931