apps: low_power: updates for testing PSM
Adjust Kconfig to define one low power mode because that is all that is needed.  The low power modes are not tied to the SOC but instead, give the app the ability to do any necessary tasks when entering or exiting power save mode.

If logging is not turned on use printk for logging messages in lte.c.

If HL7800 modem driver is disabled hold it in reset when main app thread runs.