low_power: PSM improvements
Add logic captures for reference as a benchmark.
PSM does not allow you to keep a TCP connection alive.
When using PSM, socket response data can take time to
arrive because the HL7800 has to wake up and reconfig the
IP connection and sockets, before sending out the data.

bug 17835